requirements.txt
aiohttp==3.9.0b0
aiosignal==1.3.1
attrs==23.1.0
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
discord.py==2.3.2
frozenlist==1.4.0
idna==3.4
multidict==6.0.4
python-dotenv==1.0.0
requests==2.31.0
soupsieve==2.5
urllib3==2.2.1
wheel==0.41.2
yarl==1.9.2
